requirements.txt
ursina
ultralytics
opencv-python
numpy
panda3d-gltf